pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "gradio>=5.38.2",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "mcp[cli]>=1.12.2",
    "pandas>=2.3.1",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers[torch]>=4.54.0",
    "unstructured[pdf]>=0.18.11",
]
